a zzukBot CC for shadow priests
This is set up for priests past level 40 that have respecced into
shadowform.
Wanding is still an integral part of this CC, I personally made the
switch after having had enough points for Shadowform and 5/5 Wand Spec.

I will add the pre-40 non Shadowform version of this once I'm done
cleaning up the CC since it is an absolute mess right now. (Even more
than this one, yeah.)

This does not yet work flawlessly, but after some checking and watching
I believe it to work well enough to be released for now.

I'm always open to suggestions or any kind of feedback, so do not shy
away from that.